+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 0.1.3
+- Fix: Increment pay index count for non zaps [(@denis2342)](https://github.com/denis2342) 
+- Improvement: Add option to config for path of last pay index tip
+
 ## 0.1.2 
 - Improvement: Add preimage tag to zap note
 
 ## 0.1.1
 - Improvement: Save paid invoice index tip to file. Restart from saved tip
-- Fix: Use invoice description for zap request description tag (@denis2342)
+- Fix: Use invoice description for zap request description tag [(@denis2342)](https://github.com/denis2342)